+++ enigma/utils/strings.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 import re
 from typing import Any, Union, List
 
@@ -7,23 +8,11 @@ chars = {
 
 
 def strip_emoji(text: str) -> str:
-    """Removes emoji from text.
-
-    :param text: Text from which emojis should be removed.
-    :return: Text without emojis.
-    """
     re_emoji = re.compile(u'([\U00002600-\U000027BF])|([\U0001f300-\U0001f64F])|([\U0001f680-\U0001f6FF])')
     return re_emoji.sub(r'', text)
 
 
 def fixed_width(text: str, width: int = 20, remove_emoji: bool = True):
-    """Shortens or extends string to exact width. Optionally removes emojis.
-
-    :param text: Text to shrink or extend.
-    :param width: Fixed text width.
-    :param remove_emoji: Tells if emojis should be removed or not.
-    :return: Formatted text with fixed width and optionally without emojis.
-    """
     if remove_emoji is True:
         temp = text
         text = strip_emoji(temp)
@@ -39,12 +28,6 @@ def fixed_width(text: str, width: int = 20, remove_emoji: bool = True):
 
 
 def separate_value(value: int, separator: str = ' ') -> str:
-    """Adds separator to big integers.
-
-    :param value: Value to format.
-    :param separator: Separator.
-    :return: Formatted value.
-    """
     result = ''
     n = str(value)
     for c in range(1, len(n) + 1):
@@ -55,11 +38,6 @@ def separate_value(value: int, separator: str = ' ') -> str:
 
 
 def upper_name(name: str):
-    """Makes first letters upper case. Not like CamelStyle but Every Single Word With Spaces.
-
-    :param name: Text to capitalize.
-    :return: Formatted text.
-    """
     if ' ' not in name:
         return f'{name[0].upper()}{name[1:]}'
     else:
@@ -80,11 +58,6 @@ def upper_name(name: str):
 
 
 def safe_lower(value: Union[str, Any]) -> Union[str, Any]:
-    """Saves value as lowercase only if it's string.
-
-    :param value: Text or Any to make lowercase.
-    :return: Lowercase text.
-    """
     if type(value) is str:
         return value.lower()
         pass
@@ -93,12 +66,6 @@ def safe_lower(value: Union[str, Any]) -> Union[str, Any]:
 
 
 def sort_nested_list(nested_list: List[List], key_pos: int = 0) -> List[List]:
-    """Sorts list by key in nested list.
-
-    :param nested_list: Nested list.
-    :param key_pos: Sorting key position.
-    :return: Sorted nested list.
-    """
     nested_list.sort(key=lambda x: safe_lower(x[key_pos]))
     return nested_list
 
@@ -108,4 +75,4 @@ def number_suffix(num: int):
 
 
 def f_btc(num: int) -> str:
-    return f'{round(num*0.0001, 4)} {chars["bitcoin"]}'
+    return f'{round(num * 0.0001, 4)} {chars["bitcoin"]}'