+++ docs/CHANGELOG.md
@@ -1,6 +1,19 @@
 # Enigma Bot changelog
 
 ## [Unreleased]
+
+## Released
+### [v0.7] - 08.02.2021
+#### Added
+- `spotify` command.
+#### Fixed
+- `ping` commands was showing negative values.
+- `giveaway` command embed.
+#### Changed
+- `suggest` command aliases and description.
+- Settings are now more Python-friendly.
+
+### [v0.6] - 04.02.2021
 #### Added
 - `unban` command.
 - `servers` command.
@@ -24,8 +37,6 @@
 #### Removed
 - `error` command.
 
-
-## Released
 ### [v0.5] - 03.02.2021
 #### Added
 - `suggest` command.
@@ -97,6 +108,8 @@
 
 
 [Unreleased]: https://github.com/AnonymousX86/Enigma-Bot/compare/deploy...master
+[v0.7]: https://github.com/AnonymousX86/Enigma-Bot/releases/tag/v0.7
+[v0.6]: https://github.com/AnonymousX86/Enigma-Bot/releases/tag/v0.6
 [v0.5]: https://github.com/AnonymousX86/Enigma-Bot/releases/tag/v0.5
 [v0.4]: https://github.com/AnonymousX86/Enigma-Bot/releases/tag/v0.4
 [v0.3]: https://github.com/AnonymousX86/Enigma-Bot/releases/tag/v0.3